+++ src/simulation/setup.py
@@ -4,9 +4,11 @@ with open("README.md", "r", encoding="utf-8") as fh:
     long_description = fh.read()
 
 setuptools.setup(
-    name="pybullet_controllers",
+    name="simulation",
     version="0.1",
-    description="This package implements controllers for the PyBullet simulation",
+    author="Dominic Reber",
+    author_email="dominic.reber@epfl.ch",
+    description="This package implements the PyBullet simulation",
     long_description=long_description,
     long_description_content_type="text/markdown",
     # url="https://github.com/pypa/sampleproject",
