+++ src/simulation/simulation/worlds/empty_world.py
